Basic implementation of a balancing operation and item centring.
When an item is set as the centre of the gallery, the other items are re-balanced so an equal (give or take one) number of items appear to left and right.

Currently it doesn't animate correctly, and balancing is always done to the left, which leads to some odd results.